moving generate soap summary to utils
